✨🎪 Update application next steps
