+++ apps/show/.env
@@ -1,5 +1,6 @@
 ANIMEAUX_URL="http://animeaux.localhost"
 APPLICATION_TOKEN_ADMIN="admin-token"
+APPLICATION_VALIDATION_START_DATE="2025-10-01"
 CARPOOL_FACEBOOK_GROUP_URL="#carpool-facebook-group-url"
 CLOUDINARY_API_KEY="mock-api-key"
 CLOUDINARY_API_SECRET="mock-api-secret"
