+++ apps/show/src/routes/_layout.exposants.candidature.$applicationId/route.tsx
@@ -1,32 +1,10 @@
 import { ErrorPage, getErrorTitle } from "#core/data-display/error-page";
 import { createSocialMeta } from "#core/meta";
 import { getPageTitle } from "#core/page-title";
-import { notFound } from "#core/response.server";
-import { services } from "#core/services/services.server";
-import { safeParseRouteParam, zu } from "@animeaux/zod-utils";
-import type { LoaderFunctionArgs } from "@remix-run/node";
 import type { MetaFunction } from "@remix-run/react";
+import type { loader } from "./loader.server";
 import { SectionInformation } from "./section-information";
 import { SectionTitle } from "./section-title";
-
-export async function loader({ params }: LoaderFunctionArgs) {
-  if (process.env.FEATURE_FLAG_EXHIBITOR_APPLICATION_ONLINE !== "true") {
-    throw notFound();
-  }
-
-  const routeParams = safeParseRouteParam(RouteParamsSchema, params);
-
-  const application = await services.exhibitor.application.get(
-    routeParams.applicationId,
-    { select: { contactEmail: true } },
-  );
-
-  return { application };
-}
-
-const RouteParamsSchema = zu.object({
-  applicationId: zu.string().uuid(),
-});
 
 export const meta: MetaFunction<typeof loader> = ({ data }) => {
   return createSocialMeta({
